+++ README.md
@@ -6,6 +6,14 @@
 * Liviu
 * jffy (Project Manager)
 
+## Folder Structure
+
+```
+src
+ |-- components (for stateless components)
+ |-- containers (for stateful components)
+```
+
 [![Waffle.io - Columns and their card count](https://badge.waffle.io/chingu-voyage5/Bears-Team-32.svg?columns=all)](https://waffle.io/chingu-voyage5/Bears-Team-32)
 
-Add-project-description-here | Voyage-5 | chingu.io
+Spotify-clone | Voyage-5 | chingu.io
