+++ public/index.html
@@ -20,12 +20,9 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-<<<<<<< HEAD
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp"
     crossorigin="anonymous">
-=======
   <link href="https://fonts.googleapis.com/css?family=Poppins:400,700" rel="stylesheet">
->>>>>>> 21aa5388432e65ec6f46e64def445ec6b32d00fd
   <title>React App</title>
 </head>
 
@@ -46,8 +43,4 @@
     -->
 </body>
 
-<<<<<<< HEAD
 </html>
-=======
-</html>
->>>>>>> 21aa5388432e65ec6f46e64def445ec6b32d00fd
