--- src/components/Footer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import { WrapperFooter } from "./style-utils";
-
-const Footer = () => {
-   return <WrapperFooter>Footer</WrapperFooter>;
-};
-
-export default Footer;
