--- src/components/Menu.js
@@ -1,8 +1,0 @@
-import React from "react";
-import { WrapperMenu } from "./style-utils";
-
-const Menu = () => {
-   return <WrapperMenu>Menu</WrapperMenu>;
-};
-
-export default Menu;
