--- src/components/SideBar.js
@@ -1,8 +1,0 @@
-import React from "react";
-import { WrapperSideBar } from "./style-utils";
-
-const SideBar = () => {
-   return <WrapperSideBar>sidebar</WrapperSideBar>;
-};
-
-export default SideBar;
