+++ src/containers/App.css
@@ -1,49 +1,176 @@
 :root {
-   /* COLORS */
-   --color-primary: #1db954;
+  /* COLORS */
+  --color-primary: #1db954;
+  --color-grey-dark-1: #282828;
+  --color-white: #ffffff;
+  --color-black: #191414;
 
-   --color-grey-dark-1: #282828;
+  /* BOX SHADOWS */
+  --shadow-section: 0 2px 5px 0 rgba(0, 0, 0, 0.2), 0 2px 1rem 0 rgba(0, 0, 0, 0.11);
+  --shadow-heading: 0 0 0.9rem rgba(0, 0, 0, 0.5);
+  --shadow-img: 0 0 0.5rem rgba(0, 0, 0, 0.4);
+  --shadow-hover: 0 0 0.8rem rgba(0, 0, 0, 0.3), 0 0 1.2rem rgba(0, 0, 0, 0.2);
 
-   --color-white: #ffffff;
-   --color-black: #191414;
+  /* FONT */
+  --default-font-size: 1.6rem;
 
-   /* BOX SHADOWS */
-   --shadow-section: 0 2px 5px 0 rgba(0, 0, 0, 0.2),
-      0 2px 1rem 0 rgba(0, 0, 0, 0.11);
-   --shadow-heading: 0 0 0.9rem rgba(0, 0, 0, 0.5);
-   --shadow-img: 0 0 0.5rem rgba(0, 0, 0, 0.4);
-   --shadow-hover: 0 0 0.8rem rgba(0, 0, 0, 0.3), 0 0 1.2rem rgba(0, 0, 0, 0.2);
-
-   /* FONT */
-   --default-font-size: 1.6rem;
-
-   /* GRADIENTS */
-   --gradient-default: linear-gradient(
-      to right bottom,
-      var(--color-primary-light),
-      var(--color-primary-dark)
-   );
+  /* GRADIENTS */
+  --gradient-default: linear-gradient(
+    to right bottom,
+    var(--color-primary-light),
+    var(--color-primary-dark)
+  );
 }
 
 *,
 *::before,
 *::after {
-   margin: 0;
-   padding: 0;
-   box-sizing: inherit;
+  margin: 0;
+  padding: 0;
+  box-sizing: inherit;
 }
 
 html {
-   /* This defines 1 rem = 10px; 10px/16px = 62.5% */
-   font-size: 62.5%;
+  /* This defines 1 rem = 10px; 10px/16px = 62.5% */
+  font-size: 62.5%;
 }
 
 body {
-   font-family: Poppins, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-      Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
-   font-weight: 400;
-   line-height: 2;
-   background-color: var(--color-grey-dark-1);
-   box-sizing: border-box;
-   color: var(--color-white);
+  font-family: Poppins, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen-Sans, Ubuntu,
+    Cantarell, 'Helvetica Neue', sans-serif;
+  font-weight: 400;
+  line-height: 2;
+  background-color: var(--color-grey-dark-1);
+  box-sizing: border-box;
+  color: var(--color-white);
+}
+
+/* http://meyerweb.com/eric/tools/css/reset/ 
+   v2.0 | 20110126
+   License: none (public domain)
+*/
+
+html,
+body,
+div,
+span,
+applet,
+object,
+iframe,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+blockquote,
+pre,
+a,
+abbr,
+acronym,
+address,
+big,
+cite,
+code,
+del,
+dfn,
+em,
+img,
+ins,
+kbd,
+q,
+s,
+samp,
+small,
+strike,
+strong,
+sub,
+sup,
+tt,
+var,
+b,
+u,
+i,
+center,
+dl,
+dt,
+dd,
+ol,
+ul,
+li,
+fieldset,
+form,
+label,
+legend,
+table,
+caption,
+tbody,
+tfoot,
+thead,
+tr,
+th,
+td,
+article,
+aside,
+canvas,
+details,
+embed,
+figure,
+figcaption,
+footer,
+header,
+hgroup,
+menu,
+nav,
+output,
+ruby,
+section,
+summary,
+time,
+mark,
+audio,
+video {
+  margin: 0;
+  padding: 0;
+  border: 0;
+  font-size: 100%;
+  font: inherit;
+  vertical-align: baseline;
+}
+/* HTML5 display-role reset for older browsers */
+article,
+aside,
+details,
+figcaption,
+figure,
+footer,
+header,
+hgroup,
+menu,
+nav,
+section {
+  display: block;
+}
+body {
+  line-height: 1;
+}
+ol,
+ul {
+  list-style: none;
+}
+blockquote,
+q {
+  quotes: none;
+}
+blockquote:before,
+blockquote:after,
+q:before,
+q:after {
+  content: '';
+  content: none;
+}
+table {
+  border-collapse: collapse;
+  border-spacing: 0;
 }
